feat(functions-global): :triangular_flag_on_post: add doc comment & unit feature in half function
Add documentation comments for half function and add new feature to check if the given number has unit or not. If it has a unit, it will do its functionality and return the result with its unit. If it has no unit, it will do its functionality and return the number only.