feat(mq-screens-iphone): :sparkles: add ip-4 mixin with iphone-4 mixin for simplicity
Add ip-4 mixin to use it instead of iphone-4 if user want simplicity when writing the mixin name. We edit the name of the main mixin with iphone-4 instead of iphone-4-4s. It will be used for both iPhone 4 and iPhone 4S. We wrote it again with this name for only simplicity when using it. We renamed the file of these two mixin with iphone-4 instead of iphone-4-4s to be clear.