refactor(docs): :recycle: Update dev-utils functions with validations by meta namespace
Update development utils functions to include type validations for error messages using the `meta.type-of` SASS function and the `$test-environment` variable. This is essential for ensuring proper data types are passed as error messages.  The modifications are aligned with maintaining code quality and preventing potential runtime errors.  We added also sentence of `meta.type-of` function in dependencies scope with updating the version of files.