refactor(enhance): :recycle: Update `flatten` version & add dependency & enhance error handling
Set version of `flatten` function to `1.2.1`, add `meta.type-of` as a dependency, and improve error handling for invalid input in the flatten function. The `flatten` function will now return an error if the parameter is not of list type. These changes enhance the reliability and functionality of the flatten function.