docs(docs): :memo: Add reference to MDN documentation for flex-grow prop
Add reference to MDN documentation for flex-grow property. This commit adds a reference to the MDN documentation for the flex-grow property to provide additional context and resources for developers working with the flex-grow mixin. The added reference can help improve understanding and usage of the flex-grow property in the project.